worker work started, the general idea is:
worker will query website api every 30 seconds (subject to change) for new jobs that need to be randomized. the worker can have a max of 5 jobs (potentially more in future, however it is limited by cpu and ram).

once the worker has randomized it will update the status on the website to allow the user to download the modified version.